// src/packages/master/types/MasterRdsEntity.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MasterRdsEntity = {
  id: string
  cpk: string
  csk: string
  pk: string
  sk: string
  masterType: string
  masterTypeCode: string
  masterCode: string
  tenantCode: string
  seq: number
  code: string
  name: string
  version: number
  isDeleted: boolean
  createdBy: string
  createdIp: string
  createdAt: string
  updatedBy: string
  updatedIp: string
  updatedAt: string
  syncFrom?: string
  syncDate?: string
  attributes?: Record<string, any>
}